+++ admin/start/env.ts
@@ -60,4 +60,11 @@ export default await Env.create(new URL('../', import.meta.url), {
   |----------------------------------------------------------
   */
   NOMAD_API_URL: Env.schema.string.optional(),
+
+  /*
+  |----------------------------------------------------------
+  | Variables for configuring companion service ports
+  |----------------------------------------------------------
+  */
+  DOZZLE_PORT: Env.schema.number.optional(),
 })